+++ lib/axlsx/util/constants.rb
@@ -79,6 +79,9 @@ module Axlsx
   # shared strings namespace
   SHARED_STRINGS_R = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/sharedStrings"
 
+  # theme rels namespace
+  THEME_R = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/theme"
+
   # drawing rels namespace
   DRAWING_R = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/drawing"
 
@@ -133,6 +136,9 @@ module Axlsx
   # shared strings content type
   SHARED_STRINGS_CT = "application/vnd.openxmlformats-officedocument.spreadsheetml.sharedStrings+xml"
 
+  # theme content type
+  THEME_CT = "application/vnd.openxmlformats-officedocument.theme+xml"
+
   # core content type
   CORE_CT = "application/vnd.openxmlformats-package.core-properties+xml"
 
@@ -186,6 +192,9 @@ module Axlsx
 
   # shared_strings  part
   SHARED_STRINGS_PN = "sharedStrings.xml"
+
+  # theme part
+  THEME_PN = "theme/theme1.xml"
 
   # app part
   APP_PN = "docProps/app.xml"
@@ -258,6 +267,9 @@ module Axlsx
 
   # drawing validation schema
   DRAWING_XSD = "#{SCHEMA_BASE}dml-spreadsheetDrawing.xsd"
+
+  # theme validation schema
+  THEME_XSD = "#{SCHEMA_BASE}dml-main.xsd"
 
   # number format id for percentage formatting using the default formatting id.
   NUM_FMT_PERCENT = 9