+++ test/tc_helper.rb
@@ -9,9 +9,11 @@ SimpleCov.start do
 end
 
 require 'minitest/autorun'
+require 'rspec/mocks/minitest_integration'
 require 'timecop'
 require 'webmock/minitest'
 require 'axlsx'
+require 'ooxml_crypt' if RUBY_ENGINE == 'ruby'
 
 module Minitest
   class Test
@@ -36,15 +38,19 @@ module Minitest
     end
 
     def mri?
-      !jruby? && !truffleruby?
+      RUBY_ENGINE == 'ruby'
     end
 
     def jruby?
-      defined?(JRUBY_VERSION)
+      RUBY_ENGINE == 'jruby'
     end
 
     def truffleruby?
-      defined?(TruffleRuby)
+      RUBY_ENGINE == 'truffleruby'
+    end
+
+    def ooxml_crypt_available?
+      defined?(OoxmlCrypt)
     end
   end
 end